% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday_cake.R
\name{birthday_cake}
\alias{birthday_cake}
\title{A gluten-free birthday cake for Fede}
\usage{
birthday_cake()
}
\value{
character
}
\description{
Prints an ASCII art birthday cake with a random message
}
