% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/we_depend_on_you.R
\name{we_depend_on_you}
\alias{we_depend_on_you}
\title{we "depend" on you}
\usage{
we_depend_on_you(reason)
}
\arguments{
\item{reason}{a numeric equals your age representing the reasons why we appreciate you}
}
\value{
a reason why we appreciate you
}
\description{
we "depend" on you
}
\examples{
we_depend_on_you(4)

}
