% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/we_depend_on_you.R
\name{we_depend_on_you}
\alias{we_depend_on_you}
\title{we "depend" on you}
\usage{
we_depend_on_you(reason)
}
\arguments{
\item{reason}{a numeric between 1 and 41 indicating both why we appreciate you and referring to your age}
}
\value{
a character string containing the reason why we appreciate you
}
\description{
we "depend" on you
}
\examples{
we_depend_on_you(4)

}
